clean core
