+++ geotrek/core/tests/test_views.py
@@ -1,32 +1,38 @@
 import os
 import re
-from unittest import skipIf, mock
+from unittest import mock, skipIf
 
 from bs4 import BeautifulSoup
 from django.conf import settings
 from django.contrib.auth.models import Permission
+from django.contrib.gis.geos import LineString, MultiPolygon, Point, Polygon
 from django.core.cache import caches
-from django.utils.translation import gettext_lazy as _
-from django.urls import reverse
-from django.contrib.gis.geos import LineString, Point, Polygon, MultiPolygon
 from django.test import TestCase
-
+from django.urls import reverse
+from django.utils.translation import gettext_lazy as _
 from mapentity.tests.factories import UserFactory
 
-from geotrek.common.tests import CommonTest
-
-from geotrek.authent.tests.factories import PathManagerFactory, StructureFactory
 from geotrek.authent.tests.base import AuthentFixturesTest
-
-from geotrek.core.models import Path, Trail, PathSource
-from geotrek.core.filters import PathFilterSet, TrailFilterSet
-
-from geotrek.trekking.tests.factories import POIFactory, TrekFactory, ServiceFactory
+from geotrek.authent.tests.factories import PathManagerFactory, StructureFactory
+from geotrek.common.tests import CommonTest
+from geotrek.core.models import Path, PathSource, Trail
+from geotrek.core.tests.factories import (
+    ComfortFactory,
+    PathFactory,
+    StakeFactory,
+    TopologyFactory,
+    TrailFactory,
+)
 from geotrek.infrastructure.tests.factories import InfrastructureFactory
-from geotrek.signage.tests.factories import SignageFactory
 from geotrek.maintenance.tests.factories import InterventionFactory
-from geotrek.core.tests.factories import PathFactory, StakeFactory, TrailFactory, ComfortFactory, TopologyFactory
-from geotrek.zoning.tests.factories import CityFactory, DistrictFactory, RestrictedAreaFactory, RestrictedAreaTypeFactory
+from geotrek.signage.tests.factories import SignageFactory
+from geotrek.trekking.tests.factories import POIFactory, ServiceFactory, TrekFactory
+from geotrek.zoning.tests.factories import (
+    CityFactory,
+    DistrictFactory,
+    RestrictedAreaFactory,
+    RestrictedAreaTypeFactory,
+)
 
 
 @skipIf(not settings.TREKKING_TOPOLOGY_ENABLED, 'Test with dynamic segmentation only')
@@ -872,53 +878,3 @@ class RemovePathKeepTopology(TestCase):
         self.assertEqual(poi.deleted, False)
 
         self.assertAlmostEqual(1.5, poi.offset)
-
-
-class PathFilterTest(CommonTest, AuthentFixturesTest):
-    factory = PathFactory
-    filterset = PathFilterSet
-
-    def test_provider_filter_without_provider(self):
-        filter_set = PathFilterSet(data={})
-        filter_form = filter_set.form
-
-        self.assertTrue(filter_form.is_valid())
-        self.assertEqual(0, filter_set.qs.count())
-
-    def test_provider_filter_with_providers(self):
-        path1 = PathFactory.create(provider='my_provider1')
-        path2 = PathFactory.create(provider='my_provider2')
-
-        filter_set = PathFilterSet()
-        filter_form = filter_set.form
-
-        self.assertIn('<option value="my_provider1">my_provider1</option>', filter_form.as_p())
-        self.assertIn('<option value="my_provider2">my_provider2</option>', filter_form.as_p())
-
-        self.assertIn(path1, filter_set.qs)
-        self.assertIn(path2, filter_set.qs)
-
-
-class TrailFilterTest(CommonTest, AuthentFixturesTest):
-    factory = TrailFactory
-    filterset = TrailFilterSet
-
-    def test_provider_filter_without_provider(self):
-        filter_set = TrailFilterSet(data={})
-        filter_form = filter_set.form
-
-        self.assertTrue(filter_form.is_valid())
-        self.assertEqual(0, filter_set.qs.count())
-
-    def test_provider_filter_with_providers(self):
-        trail1 = TrailFactory.create(provider='my_provider1')
-        trail2 = TrailFactory.create(provider='my_provider2')
-
-        filter_set = TrailFilterSet()
-        filter_form = filter_set.form
-
-        self.assertIn('<option value="my_provider1">my_provider1</option>', filter_form.as_p())
-        self.assertIn('<option value="my_provider2">my_provider2</option>', filter_form.as_p())
-
-        self.assertIn(trail1, filter_set.qs)
-        self.assertIn(trail2, filter_set.qs)
